requirements.txt
# Web UI Framework
nicegui>=2.7.0

# Async HTTP & WebSocket
httpx>=0.27.0
websockets>=12.0
aiohttp>=3.9.0

# Protobuf
protobuf>=5.27.0

# JS Execution (备用,推荐使用 Node.js 签名服务)
# pyexecjs>=1.5.1  # 已移除,改用 Node.js 服务

# Data Processing
pandas>=2.2.0
openpyxl>=3.1.2

# Database (SQLite wrapper)
aiosqlite>=0.19.0

# Scheduling
APScheduler>=3.10.0

# Configuration
pyyaml>=6.0
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Logging
loguru>=0.7.0

# Utilities
python-dotenv>=1.0.0
